# specs/api.yaml
openapi: 3.0.3
info:
  title: Personal AI Assistant API
  version: 1.0.0
servers:
  - url: https://api.assistant.app
  - url: http://localhost:8000
...
paths:
...
  /api/v1/voice/memory:
    post:
      tags: [Voice]
      summary: Upload voice memory
      description: Upload audio for transcription, extraction, and confirmation proposal.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "200":
          description: Transcription and extracted memory proposal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryProposalResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/voice/question:
    post:
      tags: [Voice]
      summary: Upload voice question
      description: Upload audio and receive an answer based on stored memories.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
      responses:
        "200":
          description: Assistant response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/question:
    post:
      tags: [Voice]
      summary: Ask text question
      description: Submit a text question and receive an answer based on stored memories.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextQuestionRequest"
      responses:
        "200":
          description: Assistant response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /api/v1/memory:
    post:
      tags: [Memory]
      summary: Save confirmed memory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveMemoryRequest"
      responses:
        "200":
          description: Memory stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/memory/{id}:
    delete:
      tags: [Memory]
      summary: Delete memory
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Memory deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteMemoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Memory not found.
...
components:
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
      description: Optional key to make POST operations idempotent on client retries.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized.
    ValidationError:
      description: Validation error.
  schemas:
    MemoryType:
      type: string
      enum:
        - expense_event
        - inventory_event
        - loan_event
        - note
        - document
    SaveMemoryRequest:
      type: object
      required: [memory_type, raw_text, structured_data, confirmed]
      properties:
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        raw_text:
          type: string
        structured_data:
          type: object
          additionalProperties: true
        structured_data_schema_version:
          type: integer
          minimum: 1
          default: 1
          description: Version of `structured_data` schema used by the client payload.
        confirmed:
          type: boolean
          description: Must be true. Memory cannot be persisted otherwise.
      description: >
        Persistence contract: required fields in `structured_data` depend on `memory_type`.
        If missing, server must return 422 with `memory.missing_required_fields`.
    MemoryRecord:
      type: object
      required: [id, memory_type, raw_text, structured_data, structured_data_schema_version, created_at]
      properties:
        id:
          type: string
          format: uuid
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        raw_text:
          type: string
        structured_data:
          type: object
          additionalProperties: true
        structured_data_schema_version:
          type: integer
          minimum: 1
          description: Version of `structured_data` schema stored for forward-compatible evolution.
        created_at:
          type: string
          format: date-time
    MemoryProposalResponse:
      type: object
      required: [transcript, memory_type, structured_data, needs_confirmation]
      properties:
        transcript:
          type: string
        memory_type:
          $ref: "#/components/schemas/MemoryType"
        structured_data:
          type: object
          additionalProperties: true
        clarification_questions:
          type: array
          items:
            type: string
        missing_required_fields:
          type: array
          items:
            type: string
        needs_confirmation:
          type: boolean
    QuestionResponse:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
        confidence:
          type: string
          enum: [high, medium, low]
        source_memory_ids:
          type: array
          items:
            type: string
            format: uuid
    TextQuestionRequest:
      type: object
      required: [question]
      properties:
        question:
          type: string
          minLength: 1
...
    DeleteMemoryResponse:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean